3.4.10b

+++ ModProject/ModCode/ModLib/Mod/ModMaster_CommonEvent.cs
@@ -13,12 +13,17 @@ namespace ModLib.Mod
 
         public virtual void OnInitConf()
         {
+            //copy new configs to debug folder
             var orgFolder = $"{ConfHelper.GetConfFolderPath()}\\..\\..\\..\\ModProject\\ModConf\\";
-            Directory.CreateDirectory(ConfHelper.GetConfFolderPath());
-            foreach (var orgFile in Directory.GetFiles(orgFolder))
+            if (Directory.Exists(orgFolder))
             {
-                File.Copy(orgFile, ConfHelper.GetConfFilePath(Path.GetFileName(orgFile)), true);
+                Directory.CreateDirectory(ConfHelper.GetConfFolderPath());
+                foreach (var orgFile in Directory.GetFiles(orgFolder))
+                {
+                    File.Copy(orgFile, ConfHelper.GetConfFilePath(Path.GetFileName(orgFile)), true);
+                }
             }
+            //load configs
             ConfHelper.LoadCustomConf();
         }
